+++ src/operations/Rename.cpp
@@ -39,6 +39,178 @@ static void processReferences(
     }
 }
 
+static void buildResult(
+    lsp::WorkspaceEdit& result, const std::string& newName, const std::vector<Luau::Location>& locations, const TextDocument* textDocument)
+{
+    if (!contains(result.changes, textDocument->uri().toString()))
+        result.changes.insert_or_assign(textDocument->uri().toString(), std::vector<lsp::TextEdit>{});
+
+    for (const auto& location : locations)
+        result.changes.at(textDocument->uri().toString())
+            .emplace_back(lsp::TextEdit{{textDocument->convertPosition(location.begin), textDocument->convertPosition(location.end)}, newName});
+}
+
+class FindTypeParameterUsages : public Luau::AstVisitor
+{
+    Luau::AstName name;
+    bool initialNode = true;
+
+    bool visit(class Luau::AstType* node) override
+    {
+        return true;
+    }
+
+    bool visit(class Luau::AstTypeReference* node) override
+    {
+        if (node->name == name)
+            result.emplace_back(node->nameLocation);
+        initialNode = false;
+        return true;
+    }
+
+    bool visit(class Luau::AstTypeFunction* node) override
+    {
+        // Check to see the type parameter has not been redefined
+        for (auto t : node->generics)
+        {
+            if (t.name == name)
+            {
+                if (initialNode)
+                    result.emplace_back(t.location);
+                else
+                    return false;
+            }
+        }
+        for (auto t : node->genericPacks)
+        {
+            if (t.name == name)
+            {
+                if (initialNode)
+                    result.emplace_back(t.location);
+                else
+                    return false;
+            }
+        }
+        initialNode = false;
+        return true;
+    }
+
+    bool visit(class Luau::AstTypePack* node) override
+    {
+        return true;
+    }
+
+    bool visit(class Luau::AstTypePackGeneric* node) override
+    {
+        if (node->genericName == name)
+            // node location also consists of the three dots "...", so we need to remove them
+            result.emplace_back(Luau::Location{node->location.begin, {node->location.end.line, node->location.end.column - 3}});
+        initialNode = false;
+        return true;
+    }
+
+    bool visit(class Luau::AstStatTypeAlias* node) override
+    {
+        for (auto t : node->generics)
+            if (t.name == name)
+                result.emplace_back(t.location);
+        for (auto t : node->genericPacks)
+            if (t.name == name)
+                result.emplace_back(t.location);
+        initialNode = false;
+        return true;
+    }
+
+    bool visit(class Luau::AstExprFunction* node) override
+    {
+        for (auto t : node->generics)
+            if (t.name == name)
+                result.emplace_back(t.location);
+        for (auto t : node->genericPacks)
+            if (t.name == name)
+                result.emplace_back(t.location);
+        initialNode = false;
+        return true;
+    }
+
+public:
+    FindTypeParameterUsages(Luau::AstName name)
+        : name(name)
+    {
+    }
+    std::vector<Luau::Location> result;
+};
+
+
+// Determines whether the name matches a type reference in one of the provided generics
+// If so, we find the usages inside of that node and construct the rename list
+bool handleIfTypeReferenceByName(Luau::AstNode* node, Luau::AstArray<Luau::AstGenericType> generics,
+    Luau::AstArray<Luau::AstGenericTypePack> genericPacks, Luau::AstName name, lsp::WorkspaceEdit& result, const std::string& newName,
+    const TextDocument* textDocument)
+{
+    bool isTypeReference = false;
+    for (const auto t : generics)
+    {
+        if (t.name == name)
+        {
+            isTypeReference = true;
+            break;
+        }
+    }
+    for (const auto t : genericPacks)
+    {
+        if (t.name == name)
+        {
+            isTypeReference = true;
+            break;
+        }
+    }
+    if (!isTypeReference)
+        return false;
+
+    FindTypeParameterUsages visitor(name);
+    node->visit(&visitor);
+    buildResult(result, newName, visitor.result, textDocument);
+
+    return true;
+}
+
+// Determines whether the name matches a type reference in one of the provided generics
+// If so, we find the usages inside of that node and construct the rename list
+bool handleIfTypeReferenceByPosition(Luau::AstNode* node, Luau::AstArray<Luau::AstGenericType> generics,
+    Luau::AstArray<Luau::AstGenericTypePack> genericPacks, Luau::Position position, lsp::WorkspaceEdit& result, const std::string& newName,
+    const TextDocument* textDocument)
+{
+    Luau::AstName name;
+    bool isTypeReference = false;
+    for (const auto t : generics)
+    {
+        if (t.location.containsClosed(position))
+        {
+            name = t.name;
+            isTypeReference = true;
+            break;
+        }
+    }
+    for (const auto t : genericPacks)
+    {
+        if (t.location.containsClosed(position))
+        {
+            name = t.name;
+            isTypeReference = true;
+            break;
+        }
+    }
+    if (!isTypeReference)
+        return false;
+
+    FindTypeParameterUsages visitor(name);
+    node->visit(&visitor);
+    buildResult(result, newName, visitor.result, textDocument);
+
+    return true;
+}
+
 lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
 {
     // Verify the new name is valid (is an identifier)
@@ -119,6 +291,13 @@ lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
 
     if (auto typeDefinition = node->as<Luau::AstStatTypeAlias>())
     {
+        // Check to see whether the position was actually the type parameter: "S" in `type State<S> = ...`
+        if (handleIfTypeReferenceByPosition(
+                typeDefinition, typeDefinition->generics, typeDefinition->genericPacks, position, result, params.newName, textDocument))
+        {
+            return result;
+        }
+
         if (typeDefinition->exported)
         {
             // Type may potentially be used in other files, so we need to handle this globally
@@ -162,6 +341,36 @@ lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
         }
         else
         {
+            // This could potentially be a generic type - so we want to rename that if so.
+            auto ancestry = Luau::findAstAncestryOfPosition(*sourceModule, position, /* includeTypes= */ true);
+            for (auto it = ancestry.rbegin(); it != ancestry.rend(); ++it)
+            {
+                if (auto typeAlias = (*it)->as<Luau::AstStatTypeAlias>())
+                {
+                    if (handleIfTypeReferenceByName(
+                            typeAlias, typeAlias->generics, typeAlias->genericPacks, reference->name, result, params.newName, textDocument))
+                        return result;
+                    break;
+                }
+                else if (auto typeFunction = (*it)->as<Luau::AstTypeFunction>())
+                {
+                    if (handleIfTypeReferenceByName(
+                            typeFunction, typeFunction->generics, typeFunction->genericPacks, reference->name, result, params.newName, textDocument))
+                        return result;
+                }
+                else if (auto func = (*it)->as<Luau::AstExprFunction>())
+                {
+                    if (handleIfTypeReferenceByName(func, func->generics, func->genericPacks, reference->name, result, params.newName, textDocument))
+                        return result;
+                    break;
+                }
+                else if (!(*it)->asType())
+                {
+                    // No longer inside a type, so no point going further
+                    break;
+                }
+            }
+
             std::vector<lsp::TextEdit> localChanges{};
             auto references = findTypeReferences(*sourceModule, reference->name.value, std::nullopt);
             localChanges.reserve(references.size() + 1);
@@ -186,6 +395,21 @@ lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
 
             result.changes.insert_or_assign(params.textDocument.uri.toString(), localChanges);
 
+            return result;
+        }
+    }
+    else if (auto typeFunction = node->as<Luau::AstTypeFunction>())
+    {
+        if (handleIfTypeReferenceByPosition(
+                typeFunction, typeFunction->generics, typeFunction->genericPacks, position, result, params.newName, textDocument))
+        {
+            return result;
+        }
+    }
+    else if (auto func = node->as<Luau::AstExprFunction>())
+    {
+        if (handleIfTypeReferenceByPosition(func, func->generics, func->genericPacks, position, result, params.newName, textDocument))
+        {
             return result;
         }
     }